+++ src/components/Loading/Loading.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Loading } from '../index';
+import { Loading, Button } from '../index';
 
 export default {
   title: 'components/Loading',
@@ -8,13 +8,24 @@ export default {
 };
 
 const Template = (args) => {
-  return <div>
-    <Loading {...args} />
-    <h1>Loading Demo</h1>
-  </div>;
+  const [loading, setLoading] = React.useState(false);
+
+  const handleLoading = () => {
+    setLoading(true);
+    setTimeout(() => {
+      setLoading(false);
+    }, 2000);
+  }
+
+  return (
+    <div>
+      <Loading {...args} open={args.open ? args.open : loading} />
+      <Button onClick={handleLoading}>Show Loading</Button>
+    </div>
+  );
 };
 
 export const _Loading = Template.bind({});
 _Loading.args = {
-  open: true
+  open: undefined
 };